fix(product-workflow): address approval interaction review findings
